更新

+++ network/IPackate.go
@@ -1,7 +1,7 @@
 package network
 
 type IPacket interface {
-	GetHeadLen() uint32
+	GetHeadLen() uint16
 	Pack(data TransitData, isPos int64) []byte
 	Unpack(binaryData []byte) (IMessage, error, func(conn IConn))
 }
